+++ src/clone-pug-opts.ts
@@ -1,5 +1,16 @@
 import { clone } from './utils/clone'
 
+type PugOwnOpts = Pick<PugPluginOpts,
+  | 'basedir'
+  | 'compileDebug'
+  | 'debug'
+  | 'doctype'
+  | 'filters'
+  | 'globals'
+  | 'inlineRuntimeFunctions'
+  | 'pretty'
+  | 'self'> & { filename: string }
+
 // used pug options, note this list does not include 'cache' and 'name'
 const PUGPROPS = [
   'basedir',
@@ -16,13 +27,13 @@ const PUGPROPS = [
 /**
  * Retuns a deep copy of the properties filtered by an allowed keywords list
  */
-export function clonePugOpts (opts: PugPluginOpts, filename: string) {
+export function clonePugOpts (opts: PugPluginOpts, filename: string): PugOwnOpts {
 
   return PUGPROPS.reduce((o, p) => {
     if (p in opts) {
       o[p] = clone(opts[p])
     }
     return o
-  }, { filename }) as PugOwnOpts
+  }, { filename })
 
 }